fix: move event builder serialization to another file
